requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
python-dotenv
deepface
textblob
google-generativeai
opencv-python
numpy
pillow
